+++ report-generator/src/main/java/com/teamscale/report/testwise/model/TestwiseCoverageReport.java
@@ -1,5 +1,8 @@
 package com.teamscale.report.testwise.model;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,12 +19,8 @@ public class TestwiseCoverageReport {
 	/** The tests contained in the report. */
 	public final List<TestInfo> tests = new ArrayList<>();
 
-	@SuppressWarnings("unused") // Moshi might use this (TS-36140)
-	public TestwiseCoverageReport() {
-		this(false);
-	}
-
-	public TestwiseCoverageReport(boolean partial) {
+	@JsonCreator
+	public TestwiseCoverageReport(@JsonProperty("partial") boolean partial) {
 		this.partial = partial;
 	}
 }