+++ CHANGELOG.md
@@ -7,6 +7,13 @@ We use [semantic versioning](http://semver.org/):
 # Next Release
 - [feature] _teamscale-maven-plugin_: Add new execution goal to batch convert .exec files into testwise coverage report.
 
+# 32.6.0
+- [feature] Profiler logs its version on startup
+- [fix] _agent_: The agent crashed while starting on some machines with "cannot construct instances of" errors
+
+# 32.5.0
+- [feature] _teamscale-maven-plugin_: Added the `upload-coverage` goal which automatically uploads JaCoCo reports to Teamscale.
+
 # 32.4.2
 - [fix] _agent_: The agent crashed while starting on some machines with "cannot construct instances of" errors